+++ src/신근재/백트래킹_1주차/화요일/스타트와링크.java
@@ -1,0 +1,13 @@
+package 신근재.백트래킹_1주차.화요일;
+
+public class 스타트와링크 {
+    //시작 15:20
+    //종료 16:00
+    //소요 40분
+
+    public static void main(String args[]) throws Exception {
+        //코드 리팩토링 하다가 코드가 날라갔어요
+        //40분 풀다가 이건 안되겠다 싶어서 다른 거 풀었습니다
+        //주말에 다시 볼 예정
+    }
+}
